refactor folders subdomain
