+++ services/web/server/src/simcore_service_webserver/folders/_common/models.py
@@ -20,7 +20,7 @@ from models_library.utils.common_validators import (
 from models_library.workspaces import WorkspaceID
 from pydantic import BaseModel, BeforeValidator, ConfigDict, Field
 
-from .._constants import RQ_PRODUCT_KEY, RQT_USERID_KEY
+from ..._constants import RQ_PRODUCT_KEY, RQT_USERID_KEY
 
 _logger = logging.getLogger(__name__)
