+++ services/web/server/src/simcore_service_webserver/folders/_folders_service.py
@@ -16,7 +16,6 @@ from servicelib.aiohttp.application_keys import APP_FIRE_AND_FORGET_TASKS_KEY
 from servicelib.common_headers import UNDEFINED_DEFAULT_SIMCORE_USER_AGENT_VALUE
 from servicelib.utils import fire_and_forget_task
 
-from ..folders.errors import FolderValueNotPermittedError
 from ..projects.projects_api import submit_delete_project_task
 from ..users.api import get_user
 from ..workspaces.api import check_user_workspace_access
@@ -24,7 +23,8 @@ from ..workspaces.errors import (
     WorkspaceAccessForbiddenError,
     WorkspaceFolderInconsistencyError,
 )
-from . import _folders_db as folders_db
+from . import _folders_repository as folders_db
+from .errors import FolderValueNotPermittedError
 
 _logger = logging.getLogger(__name__)
