+++ services/web/server/src/simcore_service_webserver/folders/_trash_rest.py
@@ -12,8 +12,8 @@ from ..login.decorators import get_user_id, login_required
 from ..products.api import get_product_name
 from ..security.decorators import permission_required
 from . import _trash_service
-from ._exceptions_handlers import handle_plugin_requests_exceptions
-from ._models import FoldersPathParams, FolderTrashQueryParams
+from ._common.exceptions_handlers import handle_plugin_requests_exceptions
+from ._common.models import FoldersPathParams, FolderTrashQueryParams
 
 _logger = logging.getLogger(__name__)
