+++ services/web/server/src/simcore_service_webserver/folders/_workspaces_rest.py
@@ -7,9 +7,9 @@ from servicelib.aiohttp.requests_validation import parse_request_path_parameters
 from .._meta import api_version_prefix as VTAG
 from ..login.decorators import login_required
 from ..security.decorators import permission_required
-from . import _workspaces_api
-from ._exceptions_handlers import handle_plugin_requests_exceptions
-from ._models import FoldersRequestContext, _FolderWorkspacesPathParams
+from . import _workspaces_repository
+from ._common.exceptions_handlers import handle_plugin_requests_exceptions
+from ._common.models import FoldersRequestContext, _FolderWorkspacesPathParams
 
 _logger = logging.getLogger(__name__)
 
@@ -28,7 +28,7 @@ async def move_folder_to_workspace(request: web.Request):
     req_ctx = FoldersRequestContext.model_validate(request)
     path_params = parse_request_path_parameters_as(_FolderWorkspacesPathParams, request)
 
-    await _workspaces_api.move_folder_into_workspace(
+    await _workspaces_repository.move_folder_into_workspace(
         app=request.app,
         user_id=req_ctx.user_id,
         folder_id=path_params.folder_id,
