--- services/web/server/src/simcore_service_webserver/folders/api.py
@@ -1,2 +1,0 @@
-__all__: tuple[str, ...] = ()
-# nopycln: file
