+++ services/web/server/src/simcore_service_webserver/folders/plugin.py
@@ -7,7 +7,7 @@ from aiohttp import web
 from servicelib.aiohttp.application_keys import APP_SETTINGS_KEY
 from servicelib.aiohttp.application_setup import ModuleCategory, app_module_setup
 
-from . import _folders_handlers, _trash_rest, _workspaces_handlers
+from . import _folders_rest, _trash_rest, _workspaces_rest
 
 _logger = logging.getLogger(__name__)
 
@@ -23,6 +23,6 @@ def setup_folders(app: web.Application):
     assert app[APP_SETTINGS_KEY].WEBSERVER_FOLDERS  # nosec
 
     # routes
-    app.router.add_routes(_folders_handlers.routes)
+    app.router.add_routes(_folders_rest.routes)
     app.router.add_routes(_trash_rest.routes)
-    app.router.add_routes(_workspaces_handlers.routes)
+    app.router.add_routes(_workspaces_rest.routes)
