+++ services/web/server/src/simcore_service_webserver/projects/_crud_api_read.py
@@ -19,7 +19,7 @@ from simcore_postgres_database.models.projects import ProjectType
 from simcore_postgres_database.webserver_models import ProjectType as ProjectTypeDB
 
 from ..catalog.client import get_services_for_user_in_product
-from ..folders import _folders_db as folders_db
+from ..folders import _folders_repository as folders_db
 from ..workspaces._workspaces_api import check_user_workspace_access
 from . import projects_api
 from ._permalink_api import update_or_pop_permalink_in_project
