+++ services/web/server/src/simcore_service_webserver/projects/_folders_api.py
@@ -6,7 +6,7 @@ from models_library.products import ProductName
 from models_library.projects import ProjectID
 from models_library.users import UserID
 
-from ..folders import _folders_db as folders_db
+from ..folders import _folders_repository as folders_db
 from ..projects._access_rights_api import get_user_project_access_rights
 from . import _folders_db as project_to_folders_db
 from .db import APP_PROJECT_DBAPI, ProjectDBAPI
